Restrict virtual method targets to known-created types
Previously even with lazy method conversion enabled, if we had a caller
to A.f and C.f overrides A.f, we would generate a possible call to C.f
even if a C is never instantiated in any reachable method. This commit
changes that, such that we need *both* a callsite and evidence of a C
being constructed to generate the callgraph edge.