loop unwinding in static init of Java enum
Static initialization of Java enums sometimes contains extra code that loops
over the array of enum elements. This often prevents initialization of the
enumeration and leads to zero coverage. This patch counts the number of enum
elements, and explicitely unwinds all loops in <clinit> to this number + 1. This
acts as a heuristic to make static inits of enums finish.