Improvements in the code for String.toCharArray in preprocessing
We now use an intermediary string_expression for argument of
String.toCharArray.
This is necessary for the string solver to keep trace of the strings
in the case the argument was not produced by another string function
(for instance a string literal).

This also cleans up the code generated for String.toCharArray.
There were some issues because some arrays where not allocated
correctly in the generated code, this should be fixed now.