Call remove-skip before coverage instrumentation
This prevents creating a spurious extra goal due to a trivial GOTO
(one which targets the natural successor instruction, e.g. goto x; x: ...)
In the Java frontend case these were usually due to a return instruction
at the end of the function, which gets lowered to GOTO end when end is the
successor in any case.

This may cause number of tests to harmlessly decrease by up to one per
function. The actual coverage achieved should not change as the GOTO is
already part of a basic block.

The advantage of doing this is (a) our behaviour is not brittle in the face
of a GOTO program processing step that (may) call remove_skip, and (b)
the GOTO program being instrumented is closer to the one shown by
--show-goto-functions, making debugging easier and the existence of goals
more logical.