Added in parameter to correct call to gen_nondet_init
Previously, since the update_in_place parameter is optional and there is
one too few parameters, the NO_UPDATE_IN_PLACE was being interpreted as
the max_nondet_array_length (i.e. of 0) and the max_nondet_array_length
was being used as the assume_not_null (presumably true) and !nullable
was being used as create_dyn_objects. Took a guess it should be true.