Fixed checking of not contains axioms
Previously, the Skolem form the not contains axioms was checked by negating it with the Skolem function for the existential still being used. This is logically incorrect and has been fixed. In the negation of the not contains axiom (whence the existential becomes a universal qunatifier), we expand the universal as a conjunction. Additionally, `string_refinement::get` did not properly get refined strings, which it does now.